requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
exceptiongroup==1.2.2
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
ollama==0.4.7
onnxruntime==1.21.0
openai-whisper==20240930
packaging==24.2
protobuf==6.30.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
sniffio==1.3.1
sounddevice==0.5.1
SpeechRecognition==3.14.2
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
